+++ rust/operator-binary/src/main.rs
@@ -4,8 +4,9 @@
 
 use std::sync::Arc;
 
+use anyhow::anyhow;
 use clap::Parser;
-use futures::{FutureExt, StreamExt};
+use futures::{FutureExt, StreamExt, TryFutureExt};
 use hbase_controller::FULL_HBASE_CONTROLLER_NAME;
 use stackable_operator::{
     YamlSchema,
@@ -31,7 +32,10 @@ use stackable_operator::{
     utils::signal::SignalWatcher,
 };
 
-use crate::crd::{HbaseCluster, HbaseClusterVersion, v1alpha1};
+use crate::{
+    crd::{HbaseCluster, HbaseClusterVersion, v1alpha1},
+    webhooks::conversion::create_webhook_server,
+};
 
 mod config;
 mod crd;
@@ -41,6 +45,7 @@ mod kerberos;
 mod operations;
 mod product_logging;
 mod security;
+mod webhooks;
 mod zookeeper;
 
 mod built_info {
@@ -65,9 +70,9 @@ async fn main() -> anyhow::Result<()> {
                 .print_yaml_schema(built_info::PKG_VERSION, SerializeOptions::default())?;
         }
         Command::Run(RunArguments {
-            product_config,
+            operator_environment,
             watch_namespace,
-            operator_environment: _,
+            product_config,
             maintenance,
             common,
         }) => {
@@ -97,15 +102,27 @@ async fn main() -> anyhow::Result<()> {
                     .run(sigterm_watcher.handle())
                     .map(anyhow::Ok);
 
-            let product_config = product_config.load(&[
-                "deploy/config-spec/properties.yaml",
-                "/etc/stackable/hbase-operator/config-spec/properties.yaml",
-            ])?;
             let client = stackable_operator::client::initialize_operator(
                 Some(OPERATOR_NAME.to_string()),
                 &common.cluster_info,
             )
             .await?;
+
+            let webhook_server = create_webhook_server(
+                &operator_environment,
+                maintenance.disable_crd_maintenance,
+                client.as_kube_client(),
+            )
+            .await?;
+
+            let webhook_server = webhook_server
+                .run(sigterm_watcher.handle())
+                .map_err(|err| anyhow!(err).context("failed to run webhook server"));
+
+            let product_config = product_config.load(&[
+                "deploy/config-spec/properties.yaml",
+                "/etc/stackable/hbase-operator/config-spec/properties.yaml",
+            ])?;
 
             let event_recorder = Arc::new(Recorder::new(
                 client.as_kube_client(),
@@ -167,7 +184,7 @@ async fn main() -> anyhow::Result<()> {
                 )
                 .map(anyhow::Ok);
 
-            futures::try_join!(hbase_controller, eos_checker)?;
+            futures::try_join!(hbase_controller, eos_checker, webhook_server)?;
         }
     }